+++ extension/src/frontend/view/ToolBox.tsx
@@ -1,10 +1,10 @@
 import { css } from '@emotion/react'
 import {
-  CursorArrowIcon,
-  CursorTextIcon,
-  ReaderIcon,
-  StopIcon,
-} from '@radix-ui/react-icons'
+  PanelRight,
+  SquareDashedMousePointerIcon,
+  SquareIcon,
+  TextCursorIcon,
+} from 'lucide-react'
 
 import { Toolbar } from '@/components/primitives/Toolbar'
 import { Tooltip } from '@/components/primitives/Tooltip'
@@ -55,11 +55,12 @@ export function ToolBox({
         background-color: var(--studio-background);
         box-shadow: var(--studio-shadow-1);
         border: 1px solid var(--gray-6);
+        padding: var(--studio-spacing-1);
       `}
     >
       <Tooltip delayDuration={0} asChild content="Stop recording">
         <Toolbar.Button onClick={onStopRecording}>
-          <StopIcon />
+          <SquareIcon />
         </Toolbar.Button>
       </Tooltip>
       <Toolbar.Separator />
@@ -68,21 +69,25 @@ export function ToolBox({
         value={tool ?? ''}
         onValueChange={handleToolChange}
       >
-        <Tooltip delayDuration={0} asChild content="Inspect elements">
+        <Tooltip
+          delayDuration={0}
+          asChild
+          content="Pick an element to add assertions to it"
+        >
           <div>
             <Toolbar.ToggleItem value="inspect">
-              <CursorArrowIcon />
+              <SquareDashedMousePointerIcon />
             </Toolbar.ToggleItem>
           </div>
         </Tooltip>
         <Tooltip
           delayDuration={0}
           asChild
-          content="Add assertions on text content by selecting it."
+          content="Add assertions on text content by selecting it"
         >
           <div>
             <Toolbar.ToggleItem value="assert-text">
-              <CursorTextIcon />
+              <TextCursorIcon />
             </Toolbar.ToggleItem>
           </div>
         </Tooltip>
@@ -96,7 +101,7 @@ export function ToolBox({
         <Tooltip delayDuration={0} asChild content="Toggle event list">
           <div>
             <Toolbar.ToggleItem value="events">
-              <ReaderIcon />
+              <PanelRight />
             </Toolbar.ToggleItem>
           </div>
         </Tooltip>