+++ src/components/Layout/ActivityBar/ProxyStatusIndicator.tsx
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react'
-import { Link1Icon, LinkBreak1Icon, LinkNone1Icon } from '@radix-ui/react-icons'
 import { Box, Flex, IconButton, Tooltip } from '@radix-ui/themes'
+import { Link2Icon, Link2OffIcon, Unlink2Icon } from 'lucide-react'
 
 import { useProxyStatus } from '@/hooks/useProxyStatus'
 import type { ProxyStatus } from '@/types'
@@ -69,11 +69,11 @@ function ProxyStatusIcon({ status }: { status: ProxyStatus }) {
   switch (status) {
     case 'online':
     case 'unhealthy':
-      return <Link1Icon color="gray" />
+      return <Link2Icon color="gray" />
     case 'offline':
-      return <LinkBreak1Icon color="gray" />
+      return <Link2OffIcon color="gray" />
     case 'starting':
-      return <LinkNone1Icon color="gray" />
+      return <Unlink2Icon color="gray" />
     default:
       return exhaustive(status)
   }