+++ src/components/primitives/Flex.tsx
@@ -1,4 +1,5 @@
 import { css } from '@emotion/react'
+import * as Slot from '@radix-ui/react-slot'
 import { ReactNode } from 'react'
 
 const styles = css`
@@ -24,6 +25,10 @@ const styles = css`
     align-items: stretch;
   }
 
+  &[data-align='baseline'] {
+    align-items: baseline;
+  }
+
   &[data-justify='start'] {
     justify-content: flex-start;
   }
@@ -46,32 +51,42 @@ const styles = css`
 `
 
 interface FlexProps {
+  asChild?: boolean
   direction?: 'row' | 'column'
-  align?: 'start' | 'center' | 'end' | 'stretch'
+  align?: 'start' | 'center' | 'end' | 'stretch' | 'baseline'
   justify?: 'start' | 'center' | 'end' | 'between' | 'around'
   gap?: '0' | '1' | '2' | '3' | '4'
   p?: '0' | '1' | '2' | '3' | '4'
-  children: ReactNode
+  px?: '0' | '1' | '2' | '3' | '4'
+  py?: '0' | '1' | '2' | '3' | '4'
+  children?: ReactNode
 }
 
 export function Flex({
+  asChild = false,
   direction = 'row',
   align = 'center',
   justify = 'start',
-  p = '0',
+  p,
+  px,
+  py,
   gap = '0',
   children,
 }: FlexProps) {
+  const Component = asChild ? Slot.Root : 'div'
+
   return (
-    <div
+    <Component
       data-direction={direction}
       data-align={align}
       data-justify={justify}
       data-gap={gap}
       data-p={p}
+      data-px={px}
+      data-py={py}
       css={styles}
     >
       {children}
-    </div>
+    </Component>
   )
 }