+++ src/components/primitives/Label.tsx
@@ -1,21 +1,16 @@
-import { css } from '@emotion/react'
 import { ComponentProps, forwardRef } from 'react'
 
-const styles = css`
-  font-size: var(--studio-font-size-2);
+import { Text, TextProps } from './Text'
 
-  &:where([data-size='1']) {
-    font-size: var(--studio-font-size-1);
-  }
-`
-
-type LabelProps = ComponentProps<'label'> & {
-  size?: '1' | '2'
-}
+type LabelProps = ComponentProps<'label'> & TextProps
 
 export const Label = forwardRef<HTMLLabelElement, LabelProps>(function Label(
-  { size = '2', ...props },
+  { children, ...props },
   ref
 ) {
-  return <label ref={ref} css={styles} data-size={size} {...props}></label>
+  return (
+    <Text asChild weight="medium" {...props}>
+      <label ref={ref}>{children}</label>
+    </Text>
+  )
 })