+++ src/types/fuse.ts
@@ -1,5 +1,8 @@
-export interface SearchMatch {
-  indices: Readonly<Array<[number, number]>>
+import { RangeTuple } from 'fuse.js'
+
+export interface Match {
+  indices: ReadonlyArray<RangeTuple>
   value?: string
   key?: string
+  color?: string
 }