+++ src/views/Recorder/EmptyState.tsx
@@ -1,11 +1,6 @@
 import { css } from '@emotion/react'
 import { zodResolver } from '@hookform/resolvers/zod'
 import {
-  DiscIcon,
-  ExclamationTriangleIcon,
-  InfoCircledIcon,
-} from '@radix-ui/react-icons'
-import {
   Box,
   Button,
   Callout,
@@ -17,7 +12,13 @@ import {
   TextField,
   Tooltip,
 } from '@radix-ui/themes'
-import { MinusCircle } from 'lucide-react'
+import {
+  AlertTriangleIcon,
+  DiscIcon,
+  InfoIcon,
+  MinusCircle,
+  TriangleAlertIcon,
+} from 'lucide-react'
 import { ReactNode } from 'react'
 import { useForm } from 'react-hook-form'
 import { useLocalStorage } from 'react-use'
@@ -148,7 +149,10 @@ export function EmptyState({ isLoading, onStart }: EmptyStateProps) {
                     border-top-left-radius: 0;
                   `}
                 >
-                  {isLoading ? <Spinner /> : <DiscIcon />} Start recording
+                  <Spinner loading={isLoading}>
+                    <DiscIcon />
+                  </Spinner>
+                  Start recording
                 </Button>
               </Flex>
             </FieldGroup>
@@ -204,7 +208,7 @@ function BrowserEventsSection({ children }: BrowserEventsSectionProps) {
               </>
             }
           >
-            <InfoCircledIcon />
+            <InfoIcon />
           </Tooltip>
         </Flex>
       </Text>
@@ -239,7 +243,7 @@ function WarningMessage({
     return (
       <Callout.Root>
         <Callout.Icon>
-          <ExclamationTriangleIcon />
+          <TriangleAlertIcon />
         </Callout.Icon>
         <Callout.Text>
           <strong>Supported browser not found</strong>
@@ -260,7 +264,7 @@ function WarningMessage({
     return (
       <Callout.Root>
         <Callout.Icon>
-          <ExclamationTriangleIcon />
+          <AlertTriangleIcon />
         </Callout.Icon>
         <Callout.Text>
           <strong>SSL/TLS certificate validation will be skipped</strong>
@@ -280,7 +284,7 @@ function WarningMessage({
     return (
       <Callout.Root>
         <Callout.Icon>
-          <ExclamationTriangleIcon />
+          <TriangleAlertIcon />
         </Callout.Icon>
         <Callout.Text>
           <strong>Proxy is offline</strong>